Added Suggested changes. * added authorization on comments and likes. * Model code reviewed. * Failure handling in controllers. * Added indexing in stories table.

+++ app/controllers/comments_controller.rb
@@ -9,28 +9,42 @@ class CommentsController < ApplicationController
 
   def create
     @comment = current_user.comments.new(comment_params)
+    authorize @comment
     if @comment.save
       respond_to do |format|
         format.js
       end
+    else
+      flash.alert = 'Error in commenting.'
+      format.js
     end
   end
 
-  def edit; end
+  def edit
+    @post = Post.find_by(params[:post_id])
+    render 'edit'
+  end
 
   def update
+    authorize @comment
     if @comment.update(comment_params)
+      flash.alert = 'Updated successfully.'
       redirect_to authenticated_root_path
     else
+      flash.alert = 'Something Went Wrong.'
       render 'edit'
     end
   end
 
   def destroy
+    authorize @comment
     if @comment.destroy
       respond_to do |format|
         format.js
       end
+    else
+      flash.alert = 'Something Went Wrong.'
+      format.js
     end
   end
