+++ app/controllers/likes_controller.rb
@@ -5,25 +5,31 @@ class LikesController < ApplicationController
 
   def create
     @like = current_user.likes.new(like_params)
+    authorize @like
     if @like.save
       respond_to do |format|
         format.js
       end
+    else
+      flash.alert = 'Something went wrong.'
     end
   end
 
   def destroy
+    authorize @like
     if @like.destroy
       respond_to do |format|
         format.js
       end
+    else
+      flash.alert = 'Something went wrong.'
     end
   end
 
   private
 
   def like_params
-    params.require(:like).permit(:post_id)
+    params.permit :post_id
   end
 
   def set_like
